# cmd/mcp-server/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST access to the Safecast radiation monitoring dataset via simplemap.safecast.org — 200M+ measurements from citizen scientists worldwide. All data is CC0-licensed and read-only. Powered by PostgreSQL+PostGIS.",
        "title": "Safecast Map API",
        "contact": {
            "name": "Safecast",
            "url": "https://safecast.org"
        },
        "license": {
            "name": "CC0 1.0 Universal",
            "url": "https://creativecommons.org/publicdomain/zero/1.0/"
        },
        "version": "1.0"
    },
    "host": "simplemap.safecast.org",
    "basePath": "/api",
    "paths": {
        "/area": {
            "get": {
                "description": "Returns historical bGeigie measurements inside the specified geographic rectangle. Falls back to Safecast REST API if no database is configured.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "historical"
                ],
                "summary": "Find radiation measurements within a bounding box",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Southern boundary latitude (-90 to 90)",
                        "name": "min_lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Northern boundary latitude (-90 to 90)",
                        "name": "max_lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Western boundary longitude (-180 to 180)",
                        "name": "min_lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Eastern boundary longitude (-180 to 180)",
                        "name": "max_lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results (1 to 10000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Measurements with count, bbox, and source",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid or missing parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/extreme": {
            "get": {
                "description": "Find the highest or lowest radiation readings in the database with full location details. Supports excluding anomalous sources.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Find extreme radiation readings",
                "parameters": [
                    {
                        "type": "string",
                        "default": "highest",
                        "description": "Direction: 'highest' or 'lowest'",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": -90,
                        "description": "Southern boundary for geographic filter",
                        "name": "min_lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": 90,
                        "description": "Northern boundary for geographic filter",
                        "name": "max_lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": -180,
                        "description": "Western boundary for geographic filter",
                        "name": "min_lon",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": 180,
                        "description": "Eastern boundary for geographic filter",
                        "name": "max_lon",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated device IDs to exclude (e.g., 'bGeigie-2113,bGeigie-456')",
                        "name": "exclude_devices",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JSON array of bounding boxes to exclude (e.g., '[{\\",
                        "name": "exclude_areas",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Extreme readings with location details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/radiation": {
            "get": {
                "description": "Returns historical bGeigie measurements within a radius of the given coordinates, sorted by most recent. Uses PostGIS spatial index for fast queries. Falls back to Safecast REST API if no database is configured.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "historical"
                ],
                "summary": "Find radiation measurements near a location",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Latitude in decimal degrees (-90 to 90)",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Longitude in decimal degrees (-180 to 180)",
                        "name": "lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "default": 1500,
                        "description": "Search radius in meters (25 to 50000)",
                        "name": "radius_m",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Maximum number of results (1 to 10000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Radiation measurements with count, source, and query metadata",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/sensors": {
            "get": {
                "description": "Lists active fixed radiation sensors (Pointcast, Solarcast, bGeigieZen, etc.) with their location, type, and last reading timestamp. Requires database connection.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "realtime"
                ],
                "summary": "Discover active fixed sensors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by sensor type (e.g. Pointcast, Solarcast, bGeigieZen)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": -90,
                        "description": "Southern boundary latitude",
                        "name": "min_lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": 90,
                        "description": "Northern boundary latitude",
                        "name": "max_lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": -180,
                        "description": "Western boundary longitude",
                        "name": "min_lon",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "default": 180,
                        "description": "Eastern boundary longitude",
                        "name": "max_lon",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of sensors (1 to 1000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sensor list with locations and last reading times",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Database unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}